requirements.backup.before_merge.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
numpy==1.25.2
psycopg2-binary==2.9.7
